Adds redone reader
